*descrições ^apt-hooks ^liquorix

+++ src/scripts/core/liquorix.sh
@@ -11,8 +11,10 @@ ICON_REPL='auto-cpufreq'
 # shellcheck disable=SC1090
 source <(curl -sSL https://estudio-debian.github.io/static/functions.sh)
 local_papirus_icon
-cd /tmp
-rm -fr /tmp/auto-cpufreq
-git clone https://github.com/AdnanHodzic/auto-cpufreq
-cd auto-cpufreq
-printf i|sudo ./auto-cpufreq-installer
+if ls /sys/class/power_supply/BAT* >/dev/null 2>&1; then
+    cd /tmp
+    rm -fr /tmp/auto-cpufreq
+    git clone https://github.com/AdnanHodzic/auto-cpufreq
+    cd auto-cpufreq
+    printf i|sudo ./auto-cpufreq-installer
+fi
